chore: refactor after discussion
